Agregado types para la definicion de pdfmake en typescript. Solventado error de SuperClass para hacer build. Separada funcion de configuracion de QuillEditor de la funcion del pdfMake.
