+++ package.json
@@ -12,6 +12,7 @@
     "node": "18"
   },
   "dependencies": {
+    "@types/pdfmake": "^0.2.8",
     "@vuepic/vue-datepicker": "^7.4.1",
     "@vueup/vue-quill": "^1.2.0",
     "axios": "^0.21.1",
