+++ src/App.vue
@@ -1,7 +1,6 @@
 <script setup lang="ts">
 import { useTheme } from 'vuetify'
 const theme = useTheme()
-
 localStorage.getItem("brunaTheme") == 'darkTheme'
   ? theme.global.name.value = 'darkTheme'
   : theme.global.name.value = 'lightTheme'
