+++ src/main.ts
@@ -1,9 +1,7 @@
 import { createApp } from 'vue'
 import router from "./router";
 import App from "./App.vue";
-import { QuillConfig } from './funciones/utilidades'
 
-QuillConfig()
 // Vuetify
 import "vuetify/styles";
 import { createVuetify } from "vuetify";
